// src/app/features/real-pnl-query/models/real-pnl-query-detail-response.model.ts
export interface RealPnlQueryDetailResponse {
  tdate: string;
  ttypename: string;
  stock: string;
  stocknm: string;
  qty: string;
  cqty: string;
  price: string;
  mamt: string;
  bcost: string;
  ttype: string;
  dseq: string;
  ccramt: string;
  cdnamt: string;
  cgtamt: string;
  fee: string;
  tax: string;
  interest: string;
  dbfee: string;
  adjdate: string;
  currnm: string;
  divamt: string;
  wtype: string;
  ioflag: string;
  seqno: string;
  uuid: string;
  isShowBtn: string;
}