// src/app/features/real-pnl-query/models/real-pnl-query-request.model.ts
export interface RealPnlQueryRequest {
  APISERVER: string;
  sid: string;
  sip: string;
  Invscode: string;
  comp: string;
  bhno: string;
  cseq: string;
  stock: string;
  stockName: string;
  ttype: string;
  action: string;
  action1: string;
  action4: string;
  currency: string;
  bdate: string;
  edate: string;
}